backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1

# Vector Database
chromadb==0.4.18

# Document Processing
PyMuPDF==1.23.8
python-docx==1.1.0
openpyxl==3.1.2
pdfplumber==0.10.3
markdown==3.5.1
beautifulsoup4==4.12.2
python-magic==0.4.27

# AI/ML
torch==2.1.1
transformers==4.36.0
sentence-transformers==2.2.2
accelerate==0.25.0
bitsandbytes==0.41.3
peft==0.7.1
datasets==2.15.0

# LangChain
langchain==0.0.350
langchain-community==0.0.5

# OpenAI
openai==1.5.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
aiofiles==23.2.1
httpx==0.25.2
tenacity==8.2.3

# Monitoring
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Model Serving (Optional)
# vllm==0.2.6  # Uncomment for production GPU deployment

